refactor: rename ButtonLinkNav to AboutUsNavButton
