+++ src/components/GenericPagesComponents/AboutUsNavButton/AboutUsNavButton.stories.tsx
@@ -1,6 +1,6 @@
 import { ComponentStory, ComponentMeta } from "@storybook/react";
 
-import Component from "./ButtonLinkNav";
+import Component from "./AboutUsNavButton";
 
 export default {
   component: Component,
@@ -11,9 +11,9 @@ const Template: ComponentStory<typeof Component> = (args) => (
   <Component {...args} />
 );
 
-export const ButtonLinkNav = Template.bind({});
+export const AboutUsNavButton = Template.bind({});
 
-ButtonLinkNav.args = {
+AboutUsNavButton.args = {
   buttons: [
     { label: "Who we are", href: "/first-one" },
     { label: "Board", href: "/second-one" },
