+++ src/components/pages/AboutUs/AboutUsSummaryCard.tsx
@@ -3,7 +3,7 @@ import { FC } from "react";
 import aboutNavLinks from "../../../browser-lib/fixtures/aboutNav";
 import SummaryCard, { SummaryCardProps } from "../../Card/SummaryCard";
 
-import ButtonLinkNav from "@/components/GenericPagesComponents/ButtonLinkNav/ButtonLinkNav";
+import AboutUsNavButton from "@/components/GenericPagesComponents/AboutUsNavButton";
 
 type AboutUsSummaryCardProps = Pick<
   SummaryCardProps,
@@ -26,7 +26,7 @@ const AboutUsSummaryCard: FC<AboutUsSummaryCardProps> = (props) => {
         $minWidth: 166,
       }}
     >
-      <ButtonLinkNav $mt={36} buttons={aboutNavLinks} ariaLabel="about us" />
+      <AboutUsNavButton $mt={36} buttons={aboutNavLinks} ariaLabel="about us" />
     </SummaryCard>
   );
 };
